data_pipeline/annotation_tool/requirements.txt
streamlit
pandas
boto3
python-dotenv
pillow
